+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 1.36.0 [2023-05-26]
 
 ## New Features