+++ NEWS.md
@@ -4,7 +4,7 @@
 
  * `makeClusterPSOCK(nworkers)` gained protection against setting up
    too many localhost workers relative to number of available CPU
-   cores. If `nworkers / availableCores()` greater than 1.0 (100%),
+   cores. If `nworkers / availableCores()` is greater than 1.0 (100%),
    then a warning is produced. If greater than 3.0 (300%), an error is
    produced.  These limits can be configured by R option
    `parallelly.maxWorkers.localhost`. These checks are skipped if