+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version (development version)
+# Version 1.37.0 [2024-02-14]
 
 ## New Features